Create timer implementation

+++ Assets/Scripts/MoveManagement/MoveData.cs
@@ -5,7 +5,7 @@ namespace MoveManagement
     [Serializable]
     public struct MoveData
     {
-        public float duration;
+        public int duration;
         public int moveAmount;
     }
 }
